requirements.txt
asgiref==3.7.2
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
fortnum==0.0.9
future==0.18.3
importlib-metadata==6.8.0
Markdown==3.4.4
openpyxl==3.0.10
openpyxl-templates==0.2.5
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.16.2
